Moved supplier service into its own file
refactored test
fixed bug
comments on test
